fastapi/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.19.0
requests==2.31.0
sqlalchemy==2.0.29
alembic==1.13.1
PyMySQL==1.1.0
pillow==10.3.0
Jinja2==3.1.3
boto3==1.34.93